requirements-dev.txt
# Основные зависимости
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
Flask-Migrate>=4.0.0
Flask-Admin>=1.6.0
WTForms>=3.0.0
Werkzeug>=2.3.0
python-dotenv>=1.0.0
Pillow>=9.0.0
email-validator>=2.0.0
bcrypt>=4.0.0
python-slugify>=8.0.0
markdown>=3.4.0
bleach>=6.0.0
openai>=1.0.0
requests>=2.31.0
schedule>=1.2.0
faker>=20.0.0
textstat>=0.7.0
nltk>=3.8.0
transformers>=4.30.0
torch>=2.0.0
psutil>=5.9.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
pymorphy2>=0.9.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Зависимости для разработки и тестирования
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
coverage>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Зависимости для продакшена
gunicorn>=21.0.0
whitenoise>=6.0.0
redis>=4.5.0
celery>=5.3.0